+++ packages/dartantic_ai/lib/src/providers/openai_provider_base.dart
@@ -73,6 +73,7 @@ abstract class OpenAIProviderBase<
     String? name,
     List<Tool>? tools,
     TMediaOptions? options,
+    List<String>? mimeTypes,
   }) {
     throw UnsupportedError('$displayName does not support media generation.');
   }