+++ packages/dartantic_ai/test/edge_cases_test.dart
@@ -11,7 +11,7 @@
 import 'package:dartantic_ai/dartantic_ai.dart';
 import 'package:test/test.dart';
 
-import 'test_helpers/run_provider_test.dart';
+import 'test_helpers/run_provider.dart';
 import 'test_tools.dart';
 
 void main() {