+++ packages/dartantic_ai/test/embeddings_test.dart
@@ -16,7 +16,7 @@ import 'dart:math' as math;
 import 'package:dartantic_ai/dartantic_ai.dart';
 import 'package:test/test.dart';
 
-import 'test_helpers/run_provider_test.dart';
+import 'test_helpers/run_provider.dart';
 
 void main() {
   group('Embeddings', () {