+++ packages/dartantic_ai/test/empty_after_tools_guard_test.dart
@@ -2,7 +2,7 @@ import 'package:dartantic_ai/dartantic_ai.dart';
 
 import 'package:test/test.dart';
 
-import 'test_helpers/run_provider_test.dart';
+import 'test_helpers/run_provider.dart';
 
 class DummyProvider
     extends
@@ -50,6 +50,7 @@ class DummyProvider
     String? name,
     List<Tool>? tools,
     MediaGenerationModelOptions? options,
+    List<String>? mimeTypes,
   }) =>
       throw UnsupportedError('Media generation not supported in DummyProvider');
 }
@@ -153,6 +154,7 @@ class WrapperProvider
     String? name,
     List<Tool>? tools,
     MediaGenerationModelOptions? options,
+    List<String>? mimeTypes,
   }) => throw UnsupportedError('Media not supported in WrapperProvider');
 }