+++ packages/dartantic_ai/test/thinking_consolidation_test.dart
@@ -14,7 +14,7 @@
 import 'package:dartantic_ai/dartantic_ai.dart';
 import 'package:test/test.dart';
 
-import 'test_helpers/run_provider_test.dart';
+import 'test_helpers/run_provider.dart';
 
 void main() {
   group(